Spinner moved to components folder
