--- src/Pages/LandingProfile.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const LandingProfile = () => {
-  return (
-    <div>
-      <h1>Pofile Page</h1>
-    </div>
-  );
-};
-
-export default LandingProfile;
