+++ src/Pages/Profile.jsx
@@ -6,7 +6,7 @@ import { Page } from "../SharedCSS";
 import styled from "styled-components";
 import ProfileBody from "../Components/ProfileBody";
 import useUserProfile from "../Hooks/useUserProfile";
-import "../Spinner.css";
+import "../Components/Spinner.css";
 
 const Profile = () => {
   const [user, loading] = useUserProfile();
